+++ NEWS.md
@@ -1,5 +1,7 @@
 # statgenGxE 1.0.8.9000
 
+* A bug is fixed that caused the means in the stability plots to be plotted as a straight line.
+
 # statgenGxE 1.0.8
 
 * Patch release for R 4.4.0. No user visual changes.